--- app/src/Code/index.ts
@@ -1,1 +1,0 @@
-export { Code } from './Code';
