+++ lib/src/components/code/index.ts
@@ -1,0 +1,1 @@
+export { Code, type CodeProps } from './Code';
